Deleted bloat files and condensed files to make things clear. Added functionality so when team data is viewed or an immaculate grid query is searched all the questions and answers are saved
