+++ Team_Roster/app/forms.py
@@ -1,13 +1,10 @@
-from flask_wtf import FlaskForm
-from wtforms import SelectField, StringField, SubmitField
-from wtforms.validators import DataRequired
-
+from wtforms import SelectField
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, BooleanField, SubmitField
 from wtforms.validators import ValidationError, DataRequired, Email, EqualTo
 import sqlalchemy as sa
 from app import db
-from app.models import User
+from app.models import Users
 
 class FindTeam(FlaskForm):
     year = SelectField('Year',choices=[], validators=[DataRequired()],id="year-dropdown")
@@ -18,7 +15,6 @@ class FindTeam(FlaskForm):
 class ImmaculateGridInput(FlaskForm):
     url = StringField('Immaculate Grid URL',validators=[DataRequired()])
     submit = SubmitField('Calculate Solution')
-
 
 class LoginForm(FlaskForm):
     username = StringField('Username', validators=[DataRequired()])
@@ -35,13 +31,13 @@ class RegistrationForm(FlaskForm):
     submit = SubmitField('Register')
 
     def validate_username(self, username):
-        user = db.session.scalar(sa.select(User).where(
-            User.username == username.data))
+        user = db.session.scalar(sa.select(Users).where(
+            Users.u_USER == username.data))
         if user is not None:
             raise ValidationError('Please use a different username.')
 
     def validate_email(self, email):
-        user = db.session.scalar(sa.select(User).where(
-            User.email == email.data))
+        user = db.session.scalar(sa.select(Users).where(
+            Users.u_EMAIL == email.data))
         if user is not None:
             raise ValidationError('Please use a different email address.')
