+++ Team_Roster/app/static/css/style.css
@@ -13,8 +13,8 @@ input[type=submit] {
     box-sizing: border-box;
     color: #fff;
     cursor: pointer;
-    font-size: 100%;
-    width: 250px;
+    font-size: 80%;
+    width: 200px;
     height: 44px;
     line-height: 1.15;
     margin: 12px 0 0 12px;
@@ -67,7 +67,6 @@ select {
 .select+label {
     margin-top: 2rem;
 }
-
 
 table {
     display: table;
@@ -458,4 +457,27 @@ main {
         padding: 5px;
         font-size: 0.6em;
     }
+}
+
+.user-list {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 20px;
+    margin-top: 20px;
+}
+.user-box {
+    background-color: #f9f9f9;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    padding: 15px;
+    width: 300px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    transition: box-shadow 0.3s ease;
+}
+.user-box:hover {
+    box-shadow: 0 8px 12px rgba(0, 0, 0, 0.2);
+}
+.user-box p {
+    margin: 5px 0;
+    color: #333;
 }
