+++ Team_Roster/app/templates/login.html
@@ -20,6 +20,6 @@
         </p>
         <p>{{ form.remember_me() }} {{ form.remember_me.label }}</p>
         <p>{{ form.submit() }}</p>
-        <p>New User? <a href="{{ url_for('register') }}">Click to Register!</a></p>
+        <p>New User? <a href="{{ url_for('main.register') }}">Click to Register!</a></p>
     </form>
 {% endblock %}
