--- Team_Roster/immaculateGridCalculations/webScraper.py
@@ -1,6 +1,0 @@
-
-def scrapeImmaculateGridQuestions(url):
-    #questions = ["Detroit Tigers","Only One Team","≤ 3.00 ERA Career","Gold Glove", "Played First Base min. 1 game", "Houston Astros"]
-    questions = ["San Diego Padres","Hall of Fame","Played Shortstop min. 1 game","Texas Rangers","Boston Red Sox","Atlanta Braves"]
-    questions = ["San Diego Padres","Boston Red Sox","Atlanta Braves","≤ 3.00 ERA Season","100+ Run Season","Save Career"]
-    return questions
